requirements.txt
fastapi[standard]
pywa
pywa[fastapi]
python-dotenv
